back_end/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
dnspython==2.6.0
exceptiongroup==1.2.0
fastapi==0.109.2
geographiclib==2.0
geopy==2.4.1
h11==0.14.0
idna==3.6
mangum==0.17.0
motor==3.3.2
passlib==1.7.4
pydantic==2.6.1
pydantic_core==2.16.2
PyJWT==2.8.0
pymongo==4.6.1
python-dotenv==1.0.1
sniffio==1.3.0
starlette==0.36.3
typing_extensions==4.9.0
uvicorn==0.27.1
